Added "safe navigation operator" to component templates
---
@Safe navigation operator(?):
This is an operator that guards the template from "null" and "undefined" values on property paths of an object.
For example, if you specify an interpolation of the following: {{events.location.city}} and the "location" attribute
might be null, you can make the interpolation don't throw exceptions for "can not get value 'city' of 'null'" by
doing the following: "events.location?.city" (added the "?").

PRETTY GOD DAMN COOL, RIGHT?