Usage of "safe navigation operator" along with *ngIf
-----
The "safe navigation operator" can be used to avoid DOM parsing exceptions, as explained on the previous commit.
In this case, we are avoiding a piece of DOM to be displayed if it doesn't has a specificy property named "location":
`*ngIf="event?.location"`
If we want to display `event.location.address` but the property `event.location` is null or undefined, we would get an error.
We can gracefully this type of checking using the "SAFE fucking NAVIGATION OPERATOR(?)" instead of `if event && event.location`.